requirements.txt
arabic-reshaper==2.1.3
asgiref==3.4.1
beautifulsoup4==4.10.0
dj-database-url==0.5.0
dj-static==0.0.6
Django==3.2.9
django-bootstrap4==21.1
django-braces==1.15.0
django-stdimage==5.3.0
future==0.18.2
gunicorn==20.1.0
html5lib==1.1
Pillow==8.4.0
psycopg2-binary==2.9.2
PyPDF2==1.26.0
python-bidi==0.4.2
pytz==2021.3
reportlab==3.6.2
six==1.16.0
soupsieve==2.3.1
sqlparse==0.4.2
static3==0.7.0
webencodings==0.5.1
xhtml2pdf==0.2.5
